refactor(handlers): extract object-types.ts from intent.ts (Stage B)
Move-only: SLASH_TYPE_MAP/EVIDENCE, KNOWN_BASE_TYPES, normalize*/objectBasePath/
URL builders, stripLlmEmptyValues, class-include helpers (492 lines, pure
utilities). intent.ts imports them back and re-exports the public surface for
back-compat. No behavior change; 3672 tests green, A1 snapshots zero-diff.
